Add `ReportContext` to avoid prop drilling (#6)
Added ReportContext containing the data and the current tab state (including the setter). This will allow the children to access that info without needing to use props.

Left the ability to use props so that it's possible to use the subcomponents without the parent Report and also to do overrides if needed. This is achieved by preferring prop value and using context value as fallback.

Also added missing default exports to each component.