Generate bundled type declarations (#10)
Added dts-bundle-generator as a dev dependency to generate and use types.d.ts. For some reason it ends in an error but it does generate the correct types. Updated build command to produce correct output and added some info in README.

Also separated type imports from regular imports and fixed types for displayStrings in ReportTabs.