Adjust column widths for better distribution (#17)
Adjusted min/max widths in both columns so that the distribution is better. Total of min widths is now 550px for both tables. We should revisit these numbers when adding column resizing.

Also updated itwinui-react version to fix table bugs. And removed direct dependency on itwinui-css so that it always matches itwinui-react.